move graph builder to graph.py, not to use function hook
